+++ accounts/urls.py
@@ -11,9 +11,6 @@ urlpatterns = [
     path('logout/', auth_views.LogoutView.as_view(template_name='accounts/logged_out.html'), name='logout'),
     
     path('register/', views.register, name='register'),
-    # path('search/', views.search_reddit, name='search'),
-    path('subreddit/', views.search_reddit, name='search'),
-    path('comments/', views.comments, name='comments')
     
     # path('password_change/', auth.views.PasswordChangeView.as_view(), name="password_change"),
