+++ explorer/forms.py
@@ -1,0 +1,5 @@
+from django import forms
+
+
+class SearchForm(forms.Form):
+    search_term = forms.CharField(label='Search Reddit')
