+++ explorer/templates/explorer/reddit.html
@@ -1,11 +1,7 @@
 {% for item in hot %}
 <div>
     <div style="background: gold;">
-        <p>{{ item.title }} </p>
-
-    </div>
-    <div>
-        <p>{{ item.author }}</p>
+        <p>{{ item.title }}</p>
     </div>
 </div>
 {% endfor %}
